Add loader stub for GGJT
